ListView auto-loading now ends and says no more themes
when all themes are loaded we now display as such. we also stop making
any requests to parse.

bug fix - reaching the end of the list would endless make requests to
parse up-to about 1req/s per user, i.e. if we had 30 users at the
bottom of the list we would reach our request limit for parse.